jwt 로직 재구현

+++ src/main/java/gallery/back/art/backend/api/cart/entity/Cart.java
@@ -12,11 +12,11 @@ public class Cart {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
     @Column
-    private int memberId;
+    private Long memberId;
 
     @Column
-    private int itemId;
+    private Long itemId;
 }
