+++ src/main/java/gallery/back/art/backend/api/item/repository/ItemRepository.java
@@ -7,5 +7,5 @@ import java.util.List;
 
 public interface ItemRepository extends JpaRepository<Item, Integer> {
 
-    List<Item> findByIdIn(List<Integer> ids);
+    List<Item> findByIdIn(List<Long> ids);
 }
